requirements.txt
# Core Dependencies
streamlit==1.45.1
torch==2.2.2
google-generativeai==0.5.2

# Audio Processing
librosa==0.10.1
soundfile==0.12.1  # Required by librosa

# Image Processing
Pillow==10.3.0

# Transformers and Models
transformers==4.41.2
accelerate==0.30.1  # Optional but recommended for model inference
sentencepiece==0.2.0  # Required by some transformer models
scipy==1.11.4  # Needed for audio models in transformers

# Env Handling
python-dotenv==1.0.1

# Optional: Ollama Python client (if available via pip)
ollama==0.1.7  # Replace with the latest pip version if needed

# Git for downloading models
GitPython==3.1.40

# HTTP requests for fallback downloads
requests==2.31.0

# Miscellaneous
numpy==1.26.4
